refactor: remove events module
