+++ crates/pop-parachains/src/relay.rs
@@ -1,11 +1,13 @@
 // SPDX-License-Identifier: GPL-3.0
 
 use crate::{call, DynamicPayload, Error};
+use scale::{Decode, Encode};
 use sp_core::twox_128;
 use subxt::{
 	config::BlockHash,
 	dynamic::{self, Value},
-	ext::sp_core,
+	events::StaticEvent,
+	ext::{scale_decode::DecodeAsType, scale_encode::EncodeAsType, sp_core},
 	OnlineClient, PolkadotConfig,
 };
 
@@ -86,6 +88,17 @@ impl RelayChain {
 			_ => None,
 		}
 	}
+}
+
+#[derive(Debug, Encode, Decode, DecodeAsType, EncodeAsType)]
+#[decode_as_type(crate_path = "subxt::ext::scale_decode")]
+#[encode_as_type(crate_path = "subxt::ext::scale_encode")]
+pub struct Reserved {
+	pub para_id: u32,
+}
+impl StaticEvent for Reserved {
+	const PALLET: &'static str = "Registrar";
+	const EVENT: &'static str = "Reserved";
 }
 
 #[cfg(test)]
