+++ crates/pop-parachains/src/utils/mod.rs
@@ -1,4 +1,3 @@
 // SPDX-License-Identifier: GPL-3.0
 
-pub mod events;
 pub mod helpers;
